+++ CsvGnome/Fields/ComponentField.cs
@@ -18,7 +18,7 @@ namespace CsvGnome.Fields
         public string Command => Components.Select(c => c.Command).Aggregate((t, n) => $"{t}{n}");
         public List<Message> Summary => Components.Select(c => c.Summary).Aggregate((t, n) => t.Concat(n).ToList()).ToList();
 
-        public string GetValue(int row)
+        public string GetValue(long row)
         {
             StringBuilder sb = new StringBuilder();
