+++ CsvGnome/Writer.cs
@@ -23,7 +23,7 @@ namespace CsvGnome
         /// <param name="fields">Fields describing the csv to write.</param>
         /// <param name="pathAndFile">File to write to (will be overwritten).</param>
         /// <param name="n">Number of data lines to write (not including column line).</param>
-        public void WriteToFile(Reporter reporter, List<IField> fields, PaddedFieldFactory paddedFieldFactory, string pathAndFile, int n)
+        public void WriteToFile(Reporter reporter, List<IField> fields, PaddedFieldFactory paddedFieldFactory, string pathAndFile, long n)
         {
             // Reset the date/time to write in [date] components
             Program.UpdateTime();
